+++ src/main/java/com/kau/capstone/domain/pet/entity/Pet.java
@@ -2,7 +2,6 @@ package com.kau.capstone.domain.pet.entity;
 
 import com.kau.capstone.global.common.BaseEntity;
 import com.kau.capstone.global.common.Gender;
-import com.kau.capstone.global.common.PetType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
