+++ src/main/java/com/kau/capstone/domain/pet/util/PetMapper.java
@@ -4,7 +4,7 @@ import com.kau.capstone.domain.pet.dto.request.PetRegistRequest;
 import com.kau.capstone.domain.pet.dto.response.PetInfoResponse;
 import com.kau.capstone.domain.pet.entity.Pet;
 import com.kau.capstone.global.common.Gender;
-import com.kau.capstone.global.common.PetType;
+import com.kau.capstone.domain.pet.entity.PetType;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
